+++ fault-cli/src/proxy/protocols/http/service/mod.rs
@@ -103,6 +103,8 @@ pub async fn run_http_proxy(
 
     loop {
         tokio::select! {
+            // let's make sure we prioritize the shutdown branch
+            biased;
             _ = shutdown_rx.recv() => {
                 tracing::info!("Shutdown signal received. Stopping listener.");
                 break;