+++ fault-cli/src/proxy/protocols/mod.rs
@@ -6,3 +6,4 @@ pub mod ebpf;
 
 pub mod http;
 pub mod tcp;
+pub mod udp;